// generated/dto/add-label-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Auth API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface AddLabelDto
 */
export interface AddLabelDto {
    /**
     * 
     * @type {string}
     * @memberof AddLabelDto
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof AddLabelDto
     */
    'color': string;
}