// generated/dto/issue-label.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Auth API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { Project } from './project';

/**
 * 
 * @export
 * @interface IssueLabel
 */
export interface IssueLabel {
    /**
     * 
     * @type {string}
     * @memberof IssueLabel
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof IssueLabel
     */
    'color': string;
    /**
     * 
     * @type {Project}
     * @memberof IssueLabel
     */
    'Project': Project;
    /**
     * 
     * @type {string}
     * @memberof IssueLabel
     */
    'id': string;
    /**
     * 
     * @type {object}
     * @memberof IssueLabel
     */
    'created_at': object;
    /**
     * 
     * @type {object}
     * @memberof IssueLabel
     */
    'updated_at': object;
    /**
     * 
     * @type {Date}
     * @memberof IssueLabel
     */
    'deleted_at': Date | null;
    /**
     * 
     * @type {number}
     * @memberof IssueLabel
     */
    'version': number;
}